Split out install command
