+++ .gitignore
@@ -1,1 +1,2 @@
 ied
+node_modules
