--- scripts/fmt
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-read -p "Are you sure you want to continue? <y/N> " prompt
-if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]
-then
-	gofmt -w *.go
-else
-	exit 0
-fi
